refactor: Move validators implementation to infra layer
