+++ src/infra/http/factories/controllers/RegisterUserControllerFactory.ts
@@ -1,7 +1,7 @@
 import { Controller } from '@core/infra/Controller'
-import { CompareFieldsValidator } from '@infra/http/validation/CompareFieldsValidator'
-import { ValidatorCompositor } from '@infra/http/validation/Compositor'
-import { RequiredFieldsValidator } from '@infra/http/validation/RequiredFieldsValidator'
+import { CompareFieldsValidator } from '@infra/validation/CompareFieldsValidator'
+import { ValidatorCompositor } from '@infra/validation/Compositor'
+import { RequiredFieldsValidator } from '@infra/validation/RequiredFieldsValidator'
 import { PrismaUsersRepository } from '@modules/accounts/repositories/prisma/PrismaUsersRepository'
 import { RegisterUser } from '@modules/accounts/useCases/RegisterUser/RegisterUser'
 import { RegisterUserController } from '@modules/accounts/useCases/RegisterUser/RegisterUserController'
