+++ src/infra/validation/CompareFieldsValidator.ts
@@ -1,6 +1,6 @@
+import { Validator } from '@core/infra/Validator'
 import { Either, left, right } from '@core/logic/Either'
 import { InvalidParamError } from './errors/InvalidParamError'
-import { Validator } from './Validator'
 
 export class CompareFieldsValidator<T = any> implements Validator<T> {
   constructor(
